+++ Runtime/Utils/OldValue.cs
@@ -1,0 +1,12 @@
+﻿namespace DSystemUtils.Utils
+{
+    public readonly struct OldValue<T>
+    {
+        public readonly T Value;
+
+        public OldValue(T value)
+        {
+            Value = value;
+        }
+    }
+}